Address signer bridge review feedback
