+++ app/inpage/ts/listenContentScriptBootstrap.ts
@@ -1,0 +1,1 @@
+listenContentScript(undefined, 'content-script')
